backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
pillow
gunicorn

